Add input to inverse steps
Then use an error value for evaluating the feature in latent space.
I compared norms for perfect reconstruction and total noise, and
they have approximately the same values. Most likely that without
normalization the comparison is pointless. So first I need to normalize
output values (also that was the point of using Oja vs Hebb).